requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=1.0.0
langchain-community>=0.0.10
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
streamlit>=1.30.0
sqlalchemy>=2.0.0
google-generativeai>=0.3.0
pytest>=7.4.0
colorama>=0.4.6
